Refactor: Correct directory structure and implement TCP light control
This commit addresses several critical issues:

1.  **Directory Structure Correction:**
    - Moves all integration files (`coordinator.py`, `flashforge_tcp.py`,
      `const.py`, `__init__.py`, `sensor.py`, etc., `manifest.json`,
      `services.yaml`, `translations/`) to the repository root.
    - Removes an erroneously created nested `custom_components/`
      subdirectory that contained some of these files.
    - Ensures the most up-to-date versions of `coordinator.py` (with
      TCP light changes) and `const.py` (with `DEFAULT_MCODE_PORT`)
      are in the root.

2.  **Implement TCP M-Code Light Control:**
    - Creates `flashforge_tcp.py` with `FlashforgeTCPClient` for sending
      raw M-code commands over TCP to port 8899.
    - Modifies `coordinator.py`:
        - Integrates `FlashforgeTCPClient`.
        - Rewrites `toggle_light` to use TCP client to send
          `~M146 r<R> g<G> b<B> F0\r\n` commands for light on/off.
        - Removes obsolete JSON RPC light control code (`_send_raw_json_command`,
          `_sequence_id` for that purpose).
    - Adds `DEFAULT_MCODE_PORT = 8899` to `const.py`.

3.  **HTTP Robustness (from previous successful work):**
    - `coordinator.py`'s `_fetch_data` and `_send_command` methods
      now use `resp.json(content_type=None)` to handle misspelled
      `appliation/json` mimetypes from the printer for HTTP API calls
      on port 8898.

This resolves the incorrect file paths and implements the light control
feature using the confirmed TCP M-code commands, separate from the
HTTP polling mechanism for other sensors and status.
